[#1910] Share the FfiProposal test fixture builder
ProposalTests, SDKSynchronizerProposeSendMaxTests, and
SlipstreamSynchronizerSendMaxTests each hand-rolled their own private
FfiProposal fixture builder. Promote the general one (from
ProposalTests, which already covered multi-step proposals,
receivedOutput/priorStepOutput/priorStepChange inputs, and
ephemeral-aware change values) into Tests/TestUtils/FfiProposalFixtures.swift,
add a single-step feeRequired-only convenience for the other two
suites, and delete the three private copies. The four pre-existing
builder copies elsewhere in the test tree are out of scope and
untouched.

Also fixes Step's inputValues convenience initializer, which used to
re-assign every stored property instead of delegating to the
designated initializer via self.init(...).

Test-infrastructure motion only: every migrated assertion is
unchanged, and the full offline suite still reports 886 tests with 0
failures.
